+++ _inc/blocks/blocks-manifest.php
@@ -3443,6 +3443,18 @@ Markdown is a text formatting syntax that is converted into HTML. You can _empha
       array (
         'type' => 'string',
       ),
+      'preselectNewsletterCategories' => 
+      array (
+        'type' => 'boolean',
+        'default' => false,
+      ),
+      'selectedNewsletterCategoryIds' => 
+      array (
+        'type' => 'array',
+        'default' => 
+        array (
+        ),
+      ),
     ),
     'example' => 
     array (